Fix typehint

Closes #115

See merge request maicos-devel/maicos!273

+++ src/maicos/modules/rdfplanar.py
@@ -53,7 +53,7 @@ class RDFPlanar(PlanarBase):
         First AtomGroup.
     g2 : MDAnalysis.core.groups.AtomGroup
         Second AtomGroup.
-    rdf_bin_width : int
+    rdf_bin_width : float
         Binwidth of bins in the histogram of the RDF (Å).
     dzheight : float
         dz height of a RDF slab (Å).
